refac suggestions to it's own function
